Improvements to name resolution/reference
(1) CollectionGrammar.selectSymbol used to fail when selecting a symbol with a different casing (e.g. uppercase when it was defined as lowercase), but symbols by definition are case-insensitive.  Now it works.

(2) CollectionGrammar (and, because some aspects were recursive, Grammar itself) had a bunch of machinery for name resolution that were either obsolete, unneeded, or only ever used during QualifyNames.  This commit eliminates the unnecessary ones and moves the necessary ones into functions in QualifyNames.

(3) More thorough tests for defaulting behavior.