Added more JoinReplace tests to explore the garden path more.
Specifically, we explore:
- varying the input length for single character deletions.
- the effect of from field being longer than the two field.
- multi-character deletions.

It appears that the effect of varying the input length is not more than would
be expected.

However, any kind of character deletions whether a straight deletion or as part
of a substitution has a huge impact. It appears to be exponential in the number
of characters deleted, even with just a single such rule in play.

Tests 8d, 9c, 9e have been left commented out because their execution time is in
the range 4-52 seconds, visiting 44,000-320,000 states during generation.

Comments have been added to this set of tests documenting the current number of
states visited during generation.