Addressing some collection/default issues
There was a lingering issue about the possibility of name collisions when a collection is assigned to a default symbol.  Say you've assigned a collection to `Verbs.Default`, and that collection has a symbol `Class1`, but also the `Verbs` collection has a `Class1`.  Which one does the symbol `Verbs.Class1` refer to?

In practice, whichever was defined first survived, and the other got a warning.  But this warning never showed up anywhere!  Why not?  It turns out CollectionGrammars never got wrapped in LocatorGrammars, and since CollectionGrammars were always top-level there was never any LocatorGrammars to localize their messages.

So of course, we add locator wrapping into CreateGrammars.handleCollection() where it belongs.  But then this messes up other things -- the algorithm for QualifyNames.handleCollection() assumes collections are direct parents/children of each other, and now they're not, there are LocatorGrammars in between.

Much refactoring later, and name qualification is now a bit conceptually simpler at the algorithmic level, and the semantics of defaults has clarified somewhat.  (I mean, maybe it's a little harder for a beginner to understand, but it's not *sloppy* whereas before it was a bit sloppy.)

BEFORE: QualifyNames.handleCollection used to assign "X.Y.DEFAULT.Z" to "X.Y.Z".  This brought up the lingering issue mentioned above, that "X.DEFAULT.Y.Z" would also be assigned to "X.Y.Z".  Meanwhile, QualifyNames.handleCollection returned a new CollectionGrammar but the actual contents of it were ignored, the return was only used for message collection.

NOW: It's somewhat simpler -- everything gets assigned/resolved to its actual name including the "default".  But, in addition, referent of "default" in any collection is returned as the return value of QualifyNames.handleCollection, and thus gets assigned to whatever the collection got assigned to.

Consequences:

(1) Everything has a full name including the "defaults", like `X.Y.Default.Z.Default`, and you can use that in embeds or to generate from.  (I don't think the programmer should actually be doing this, this really isn't what the word default is really for, and I think it's valid for interfaces to hide these symbols.  But they're there underneath.)

(2) Abbreviations don't refer to collections, only to default grammars.  That is to say, if it happens that `X.Default` is a collection, `X` is NOT an alias to that collection in the sense that you can pick out symbols from it using `X.something`.  It's only a reference to X's default grammar, whatever that happens to be.  (And that in turn might be the default's default grammar, and so on.)

You can think of this like "If there's a symbol `X.Y.Default.Default.Default`, you can abbreviate it like `X.Y`."  But that's all you can do, only the ending "Default"s can be left off, not ones in the middle.

(We could, and briefly did, allow the name resolver to jump into default namespaces to resolve names in the absence of an alternative hypothesis.  But this ended up being confusing because then embeds and other symbol-related things work differently -- later symbol-related things don't have access to the rich structure that let embed resolution make this jump.  Better off just saying "That's not how abbreviation works.")

This also means you can't just ask the interpreter to generate from some symbol "Verb" in your main sheet anymore, you have to qualify it like "SheetName.Verb".

(3) Eliminating the name ambiguity also eliminates the possibility of name collision, which means that the possibility of the warning goes away.  With that warning gone, there aren't actually any messages that CollectionGrammars fire, and so they don't have to be located, so they don't have to be wrapped in LocatorGrammars, so the original motivation for this refactoring becomes moot.

But it's still good to wrap these in LocatorGrammars, someday these will fire messages again, so we might as well address this now.

(4) The recently-added check that AssignDefaults did, to check if the root was an NsGrammar and to make it one, if not, goes away.  It's never true anymore, the root is a LocatorGrammar now.  But what it did is now moot, with the new way of handling QualifyNames.

(5) We're back to there being more symbols inside CollectionGrammars that refer to the same actual grammar, which is just in general something we want to avoid.  (When a grammar is multiply represented in a collection because multiple symbols refer to it, it's transformed multiple times and independently, which can potentially lead to bugs or inefficiencies.)

This is less of a pressing issue than it used to be, though, because in practice most default symbols are going to be auto-generated and therefore embeds.  But anyway, maybe in the future we should replace CollectionGrammar's internal Dict<Grammar> with a collection type that allows multiple names for each element, to do away with this worry entirely.