Capitalization fixes
(1) A few commits ago, we switched from TstAssignment using cell.text for its symbol to using op.text.  (It was no longer true that the symbol name was identical to the cell text.)

However, the op parser was lowercasing everything for convenience.  (It didn't matter, we had never been showing the results of this parse to the programmer before.)

So now that an output of the op parser IS being shown back to the programmer, it wasn't respecting their original capitalization.

I stopped lowercasing the input text to the op parser.  This doesn't cause any direct problem (op parsing was case-insensitive already), but if consumers of the parse expect lowercase outputs (as in (2) below), they'll have to remember to lowercase them.

(2a) When making binary ops (in CreateOps.handleBinary), we were looking up ops by string, and expecting lowercase.  So now we have to lowercase the outputs.

(2b) We weren't actually testing for variant-case binary ops (like "JOIN:"), now we are.

(3) The GSuite UI now ignores any symbol that ends in "Default", because there will be an abbreviation of this name with the same semantics.

(This undoes a recent change.  It was nice for the UI to not have to know about defaulting and such, but defaulting has changed since then, and the UI has to know about it again.)