Refactoring binary op handling
A change to capitalization in the last commit messed up binary ops ("join:", "or:", "concat:") and only those.  Why?  Because we handle binary ops in a bit different way from other operators.  Instead of the op parser making different Ops/TSTs for them, they all went in BinaryOp/TstBinaryOp and we defer figuring out what exactly they denote to CreateGrammars, which looks up op.text in a special dictionary to figure out what grammar it consists of.

That made sense when TSTs were relatively heavyweight and we could avoid duplicating all the TST code for each op.  But TST classes are now very lightweight, all they have to know how to do is mapChildren, so we don't actually save a lot of code this way.  And handling them differently than normal means a potential source of issues like the one we just encountered -- the general, built-in case-insensitivity of this subsystem didn't apply to it, because it was using its own mechanism.

So this commit makes binary ops work like everything else.  Join and Or are now their own Ops/TSTs, and Concat is gone because no one ever used it for anything.