Added yet more JoinReplace tests to explore the garden path.
Some missing test cases have been added and the tests have been renumbered in a
slightly more systematic way.

Tests 8a-1 to 8a-5 and 9a to 9d demonstrate that the garden path for character
deletions occurs even for just a single rule. The problem occurs whether we are
doing a substitution that reduces the number of characters or a straight deletion.
The number of states visited in generation grows in an exponential manner as
the number of characters removed increases.

Tests 12a to 12c show that the garden path problem doesn't exist for insertions
(i.e. when the number of from characters is less the number of to characters).