Implemented EpsilonLiteral{Grammar,Expr}, Epsilon with a tape.
Experimented with it in testJoinReplace as a way to deal with the garden path
problem when characters are being deleted (i.e. # from chars > # to chars).

Tests for a single rule (8a5-test-1 to 8a5-test-4 and 9d-test-1 to 9d-test-4)
show excellent promise that adding #from-#to EpsilonLits solves the issue.

However, test 8b-3-test-1 to 8b-3-test-12 don't paint a clear picture for what
to do with rule cascades.

I had also initially implemented Pat's idea for MatchExpr of adding an EpsilonLit
to opposite tape when an EpsilonToken is encountered instead of just moving on.
This did not work, so I have commented it out. I had added tests X1 and X2, which
don't really belong in testJoinReplace, to experiment with this.