Removed maxReps/maxChars from Not, NegationGrammar and NegationExpr.
In usage where maxChars != Infinity, we now wrap the Negation Grammar/Expr in a
CountTape Grammar/Expr to limit the number of characters generated.

I also corrected CountTapeExpr.deriv to check that maxChars holds a value for
the tape in question before decrementing the value.

I will explore dropping maxExtraChars and maxCopyChars from ReplaceGrammar in a
subsequent commit.