Local priority decisions
This branch experiments with making tape priority decisions at a more local level, rather than by a global loop like before.  This doesn't *yet* solve our problem, but I suspect it's a step on the path towards it.

Previously, a single high-level object (like a Generator or PriorityExpr) maintained a sequence of tape names and queried those on a loop, like a giant clock where every tick moves on to another tape.  But we suspect that we may need to make decisions about this sequence that's hard to make at the top level, which is pretty blind to what tapes are *for*.  It doesn't know that T3 and T4 have any sort of relationship, it just knows they exist and that some other process said that T3 has got to precede T4.  If we need to skip querying T4 because T3 came back epsilon, that's not a decision the top-level loop can make.

So this branch experiments with a new way of doing priority locally, so that there's a kind of expression that says "I know you asked for a T4, but gimme a sec, I have to query T3 first."