Added some Repeat tests for repeats on multiple tapes with different # symbols.
Pat had some concern about whether a RepeatExpr needs to emit the same # symbols
on each tape it cares about.

The tests I wrote don't show any sign of any such issue. The correct results are
generated with no signs of exponential behaviour.

I also improved the output logging for the tests reporting the number of states
in testRepeat and testJoinReplace, outputting both the file name at the start
and the each test name. Output of the names and number of states can also be
turned on/off globally for the file. The additional name output is very useful
when running tests using the Mocha Test Explorer in VSCode because it appears
that all tests in all files are run even if only a single file or test has been
selected to run.

As part of the output logging improvements I enhanced the logDebug function in
util.ts to accept multiple message arguments.