Equal-length correspondences with auto-epsilon
So in the last commit, it seemed to work that equal-length input/output rules didn't cause the garden path phenomenon.  When correspondences weren't equal length, we had to pad the length with epsilon literals.

That's fine when the correspondences are literals, and we can know their length!  But what about when they're not?  E.g., if the rule is ("s"|"ts")=>"", we can't just pad the output with the appropriate number of epsilon literals.

So we have to have something that keeps track of the actual number tokens emitted from the "from" side, and allow the emission of epsilons until we reach the same number on the "to" side.  MatchCountExpr was close, but didn't do the epsilon part.

CorrespondExpr is a more narrowly-conceived MatchCountExpr (it only works in this very specific situation), but handles both the counting and the emission.  It's not symmetric, it's not that it requires them to be equal length and will emit epsilon tokens on either tape.  That ended up being quite tricky to do properly, and I gave up for now, because only one kind of inequality (toCount < fromCount) was actually causing the garden paths.  So CorrespondExpr makes sure toCount >= fromCount, and emits epsilon tokens on the toTape, but that's it.

At least in this context, this epsilon emission seems to be sufficient to get all the right answers; if we comment out the epsilon emission in RepeatExpr all these examples still work.  (But we still may need it for RepeatExpr tests themselves.)

Shorter-from rules (e.g. "a" => "AA") no longer work, but after some investigation I don't think it's a problem with correspondences.  I think it's just that PreTapeExpr needs to be a little smarter about delta-ing out.