Allow early nulling of PreTapeExpr.fromTape
In the previous commit, we found that the new priority/correspondence system was making rules like `a => AAA` not work.  My sense was that this was ultimately a problem with PreTapeExpr rather than CorrespondExpr.  Fortunately it was, and the fix was easy.

As written, PreTapeExpr was effectively requiring that its fromTape and toTape become nullable at the same time -- it could only delta out both at the same time.  And that's exactly what was happening in the rules we've been trying lately, either because they already were the same length, or CorrespondExpr forced them to be the same length.

But CorrespondExpr only does anything to rules like `aaa => A`.  It's tricky to get it to work the other way, like at the point where `a => AAA` has output the toTape:a and fromTape:A, that's a 1=>1 count.  When we're asked to emit from toTape again, how do we know whether or not to allow the emission of an epsilon?  It depends on what fromTape is going to do next, and that's looking into the future.  I can think of some directions but it's not really a path I wanted to pursue.

Instead, we can address this by relaxing that implicit requirement of PreTapeExpr that they become nullable at the same time.  Just conceptually, that's wrong.  PreTapeExpr and PriorityExpr serve the same role in different ways; at one point we imagined PreTapeExpr as eventually taking over from PriorityExpr and becoming the only priority system, like where there's no big PriorityExpr, just a chain of PreTapeExprs that in aggregate handle every tape.  That obviously can't work if PreTapeExpr.delta has the requirement that every tape null out during the same cycle!

So this commit makes a new way to "move on" when toTape is nullable but fromTape is not necessarily so.  During PreTapeExpr.deriv, if the child is nullable on toTape, we call deriv(fromTape) on that, and don't wrap the result in a PreTapeExpr.  (That's how the pre-tape system stops querying its tape, it just stops existing.  PreTapeExprs always and only query on that tape, and that tape is only ever queried by its PreTapeExpr, so to stop querying on it forever, all that's necessary is that the PreTapeExpr stop reproducing itself.)