Bugfix for counting epsilon tokens
I was testing that all the old tests still worked, and found a few remaining issues, only one of them an actual bug:

(1) testReplace test 37k was failing.  Explaining the problem is hard, but the solution is easy.

The test: the free generation from the replacement rule t1:"e" => t2:"", with vocab "eh", and a max count {t1:3,t2:1}.

The problem: Some valid results are generated (like {t1:eeh,t2:h}, others are not generated ({t1:ehe,t2:h}).  The issue is with the interaction of CorrespondExpr and CountTape.  CorrespondExpr needs the same number of emitted tokens on both tape, and if we're already done on t2, will keep emitting epsilon tokens until both are done.

This works fine for {t1:eeh,t2:h}.
  -- We emit t1:e, count is now {t1:2, t2:1}
  -- We emit t2:eps, count is still {t1:2, t2:1} because we don't decrement when it's epsilon
  -- We emit t1:e, count is now {t1:1, t2:1}
  -- We emit t2:eps, count is still {t1:1, t2:1}
  -- We emit t1:h, count is now {t1:0, t2:1}
  -- We emit t2:h, count is now {t1:0, t2:0}, everyone's satisfied, YIELD.

This doesn't work for {t1:ehe,t2:h}
  -- We emit t1:e, count is now {t1:2, t2:1}
  -- We emit t2:eps, count is still {t1:2, t2:1}
  -- We emit t1:h, count is now {t1:1, t2:1}
  -- We emit t2:h, count is now {t1:1, t2:0}
  -- We emit t1:e, count is now {t1:0, t2:0}
  -- We can't delta out because CorrespondExpr isn't yet satisfied, but we also can't deriv because CountTape won't let us deriv with t2:0 remaining.

The solution: We were forbidding that deriv too early, before we knew whether the result was epsilon or not.  It's only AFTER doing the deriv that we can determine if the result violates the remaining count -- epsilons never will because they don't decrement the remaining count.

(2) The multi-output-tape replace rules are failing, since CorrespondExpr is a two-tape-only thing (it throws an exception if you attempt otherwise) and these are correspondences between 3+ tapes.

I commented out these tests for now.  I don't think it would actually be that hard to get CorrespondExpr working with multi-tape rules... but also, the whole package of ideas containing CorrespondExpr is a solution for composing 1-tape -> 1-tape rules without garden paths.  1-tape -> 2-tape rules aren't something we can offer to the programmer at all yet, our efficient solution doesn't include those.  (In fact, our current solution excludes them.)  So anyway, there's not much call yet for writing a more complicated CorrespondExpr just to get these tests right.

(3) I had commented out the .END tapes in EpsilonExpr and NullExpr for some reason (probably was fixing something where .END was a distraction), so the tests that required it failed.  I put these back in.