Preventing infinite loop from nullable input
The interaction of RepeatExpr returning epsilon tokens, and the way PreTapeExpr handles epsilon tokens in the input tape (it doesn't query its output tape, it just returns epsilon), leads to an infinite loop, as PreTapeExpr keeps querying its input tape, gets epsilon, and tries the same thing again the next time.  Nothing changes each time, except we keep buffering the output side of the rule (but never actually querying on the outputTape to emit that output).

This commit adds a simple EpsCountExpr that doesn't allow two epsilon token returns in a row.  That short-circuits the infinite behavior of concern.  It doesn't actually prevent the multiple execution of the rule, just the situation where we repeatedly emit epsilon on the same tape in a row.