Amendment to previous bugfix
I was wrong about what the previous bugfix actually did.  It worked by limiting repeated unconditional generations to two or fewer.  (I shoulda been testing without joins to see that!)  That's a bizarre semantic restriction.  Limiting `abc & ""->D` to only generating one D in a row, that's reasonable, limiting it to two D's in a row is arbitrary.

The right number count to pass to EpsCountExpr to get ONE in a row wasn't 1, it was 0.  In other words, this should have been NoEpsExpr.  So now it is.

This still has one arbitrary and bizarre output, that it's allowed to generate infinitely at the ends of words.  Like `abDDDDDc` is forbidden, but `abcDDDDDD` is fine.  This is because, at the end there, we've delta'd out of the input tape and all that's left is a repeat of the output tape.  We should probably fix this eventually but for now I feel this is a rare special case of a rare special case, and put it on the back burner.