Restoring tests; all pass
Two notes:

(1) As mentioned before, 1-tape -> 2+-tape replacement tests are still commented out, because our current solution doesn't encompass these.

(2) The weird corner case behavior of unconditional insertion rules (where they can keep inserting forever at the end) is in the tests as "correct" for now.  It's a stupid behavior but it's not a surprise, I don't feel like dealing with it right now, and we want a warning if anything we do changes this behavior, so it's "correct" in the tests.