Restored state counting
When we moved some aspects of priority into PreTapeExpr, the resulting calculations didn't get counted as visiting a state.  State-counting was only done as part of the outermost loop, and PreTapeExpr didn't have access to that count and couldn't increment it.

This commit introduces a DerivStats object into DerivEnv.  Since the env is threaded through the recursive algorithm, PreTapeExpr has access to it.

Anything else we want to count, like (say) how many derivs/delta we call, can be added to DerivStats.

Note that, unlike a lot of recent implementation changes, DerivStats is stateful/impure (that is, incrementing it is treated as a state-changing side effect).  If you reuse a DerivEnv or DerivStats for some reason (say you're using one for multiple unit tests), it won't reset, it'll just keep incrementing.