Removing bitsets in favor of string sets
Reworking how outputs work and threaded through the system is a pretty major refactoring, and it's complicated by the fact that a lot of code is doubled or branches based on whether we're working on bitsets or strings, and bitsets also have all the Entanglement issues.

But we're not actually using bitsets for much; they're used for three things:

    (1) As a possible component of a deriv return, which was intended as an efficiency optimization.  However, we're not actually using them, and when we do they're actually slower.  (That slowness is probably due to Entanglement not being as efficient as it could, but it's not really worth pursuing much further.)

    (2) To hold the "remainder" necessary for the calculation of negation.

    (3) To hold vocabs.  But this was actually due to (1), vocabs needed to be bitsets in order to make it efficient to switch BitsetTokens between tapes.

This commit removes all everything about bitsets and entanglement, makes stringDeriv the only deriv, and handles (2) and (3) with Set<string> instead.  This ends up eliminating a lot of the complexity of tapes.ts.