Simplifying outputs
When we got into entanglement, this greatly increased the complication of outputs, to the point where we had to abandon our non-recursive implementation of toDict for the time being.

With the removal of both entanglement and bitsets, outputs and their conversion to StringDicts become very simple.  There's no longer any ambiguity, so the result is a StringDict, not StringDict[].  I've also renamed it from OutputTrie to OutputExpr, moved it into expr.ts, and renamed it toDenotation().

The general idea is that some expressions have easily calculable (e.g. non-recursively collectable) and unambiguous denotations -- EpsilonExpr, LiteralExpr, and concats where the last element (or first, if we're going the other direction) has such a denotation.  Others don't have this -- we can only gather their denotation through the generation algorithm.  The purpose of the algorithm is to turn expressions of the latter type into expressions of the former type.