Refactoring OutputExpr -> ConcatExpr
OutputExpr was really just a special case of Concat.  (Or you could think of it as a Precede, a Concat who constructs its children according to the RTL/LTR setting, although recall that we don't actually have a PrecedeExpr, it's just a smart constructor.)

So I moved toDenotation to ConcatExpr and use those instead.  The former .prev and .item (or whatever I happened to call it) are just ConcatExpr's children now.

Two important differences from before:

(1) We used to use OutputTries/OutputExprs identically regardless of parse direction.  (I.e., regardless of LTR/RTL, OutputExpr.item was always the target being added to the tree, and OutputExpr.prev was always the rest of the tree.)  This made some aspects of the algorithm easier (like you always added items identically) but I found it easy to get things backwards when finally turning things into dicts.  (E.g., in RTL we were doing a backwards traversal of a tree that itself is storing things in a backwards order, so when concatting item and prev, which direction should they come in to get the output right?)

It makes me uncomfortable using ConcatExpr like that.  In a ConcatExpr, we should really be storing things in the order they really occur in, so that the left child is on the left and the right child is on the right.  Everything else about ConcatExpr assumes that, after all.

So now that we're using ConcatExpr, we store outputs in their actual order.  When we're adding target chars to the right of the word, they're placed on the right; when left, left.  You could take the resulting expression and do whatever with it, it's a properly-constructed expression. (But actually, don't call .id on these, outputs get really deeply nested and id is recursive, so it can get exponentially slow.)

Storing outputs like this makes it much easier (for me at least) to conceptualize what's happening in generation.  You're turning an Expr into an equivalent Expr, with the same eventual denotation, but expressed in restricted set of simpler components, and in a structure we can traverse nonrecursively and in linear time.  (Or, to be precise, turning an Expr into a set of such Exprs, the union of which would have the same denotation.)

Note that, compared to before when we only cared about LTR/RTL in one place, we now have to care about it in two places:

  * Once when adding, so that we add them to the appropriate side
  * Once when traversing so that we visit nodes in the appropriate order.  (Not that we get the *wrong* answer when going in an inappropriate traversal order now, because everything's always stored in the correct order.  But the point of doing this in the first place is to eliminate tail-call recursion so that we don't risk blowing the stack; we need to eliminate tail-call recursion in the direction that actually recurses!)

(2) Note that this isn't typesafe the way OutputExpr was.  OutputExpr could only have children of particular types; items were always Literals and prevs were either Epsilons or OutputExprs.  ConcatExpr can have any Expr as either child, and not every Expr has .toDenotation() implemented.  We're relying on the generation algorithm to make well-structured and linearly-traversible outputs, but this isn't enforced at compile time through types.

We *could* genericize ConcatExpr to make everything type safe, but recursive generic types are dicey, and it would make LTR vs. RTL ugly because those would end up using different type parameters.  (And if we're threading outputs through deriv/delta in the future, that's a type we use all the time.  I really don't want deriv/delta to have different parameterizations depending on parse direction!)