Fixed a couple small bugs in testUtil in testMatchOutputs and testParseMultiple.
In testMatchOutputs, the increment used in looping could have an inappropriate
value when the outputs were excessively large, and the expected outputs were
very small.

In testParseMultiple, a completely empty dictionary for an input would result in
full generation, which is particularly problematic because testParseMultiple is
used in exactly the situation when the number of outputs may be very large. We
now throw an exception in this case since this is a programmer error.