Removed maxExtraChars, maxCopyChars and vocabBypass from ReplaceGrammar and Replace.
maxExtraChars and maxCopyChars existed to make writing tests easier, but they
complicated the code. The introduction of CountTape reduced the need to use
maxExtraChars < Infinity. Also, maxCopyChars < maxExtraChars or CountTape value
could lead to unexpected results.

Similarly, vocabBypass=false was used only in unit tests, and did not reflect
the actual language use case.

The tests in testReplace have been extensively refactored for the removal of
these three parameters. In some cases we switched to using testParseMultiple
instead of testGrammar because of the large number of outputs generated for a
useful size test.

Finally, we removed the ReplaceBypass functions from various test files as it is
no longer needed.