Unifying symbol and tape name restrictions
Before, we didn't have/enforce consistent restrictions on identifier names for symbols or tapes, except "if it parses, it works".  Since the special characters used in Op cells (where symbol names live) and Header cells (where tape names live) were different, this actually meant that the two kinds of identifiers had different rules.  This would lead to trouble down the road.  (Actually, it could already lead to trouble, since Op parsing was using the symbol rules for all identifiers in Ops, but the "replace X:" op contains a tape identifier!)

We also we cavelier about Unicode: putting Unicode characters in happened to work, but not because we thought about it, just because nothing in the parser stopped you from doing so.

Anyway, it was high time to be strict about this.  Now:

* Identifiers have to start with letters or underscore.  Unicode categories L, M, and Sk count here.
* Beyond letters and underscores, they can also contain numbers (including Unicode category N) and a limited set of symbols (@ # $ & ? + - * ^ = ' "); they just can't start with these.
* Nothing else can be in an identifier

I added tests to testHeader.ts for these, and also noticed that we weren't testing Ops in the same way.  (There just wasn't much interesting to test for.)  Now there's a testOp.ts, too.