Fixed two bugs w.r.t. Priority and CollectionGrammar/Expr.
Interpreter.prepareExpr was always wrapping our grammar in a new default
PriorityGrammar because now the targetGrammar is a CollectionGrammar, so we
need to drill down to determine if there is already a PriorityGrammar in play.
As a result, calls to Priority in unit tests were effectively being ignored.
Now we drill down to look for a PriorityGrammar if targetGrammar is a
CollectionGrammar.

Once the above issue was fixed, a second problem showed up. In generate, inside
the generation loop, we have a case where prevExpr can be either a PriorityExpr
or a CollectionExpr. For the CollectionExpr case we were failing to drill down
into the inner PrioriyExpr to access tapes.