Prepping mini parsers for context parsing
Parsing contexts for linguistic rules (e.g. #A_B#) is a bit tricky to do concisely.  A or B are regexes, and so it'd be nice to just be able to re-use the regex grammar, but the terminals of that grammar like RESERVED and UNRESERVED depend on what characters are reserved.  But that's different for regexes and contexts; in addition to the previous, contexts also reserve # and _.  (On the other hand, if A or B is in parentheses, then # and _ could be treated as unambiguous literals; # and _ are only special in contexts at the top level.)

This isn't quite possible the way we had the miniParser set up, things like RESERVED and UNRESERVED for any particular grammar took a particular set of reserved words/characters at their construction; it didn't depend on context.

In preparation for context parsing, this commit changes that, so that the reserved-word set comes from an env variable, and there's a parser that's able to swap out that env variable for a different one within its scope.  (This is very much like how renaming, recursion protection, etc. work in the rest of Gramble.)

Note, though, that this doesn't give us complete freedom to nest different kinds of grammars with different tokenizations within each other.  The nested grammar (here, regexes) must reserve a subset of the words reserved by the outer grammar (here, contexts), or else some of the things the nested grammar will be looking for won't have been tokenized.

(For example, consider if regexes thought * was an operator but contexts didn't.  When we tokenize the whole string at the beginning, according to the assumption that it's a context, things like "ab*cd" would remain single tokens, and when we're down into regexes the parser would go "Is that a *?  Nope!" and treat it as a literal.)

Fortunately, our regex special characters are, by nature, a subset of our context special characters, or else we'd need to rewrite all parsing to be character-based.  (Which is possible but rather tedious.)