Adding dot support to regexes
We didn't actually have support for dot (in the sense of "match any character") in regexes, but ultimately we need those for a variety of things.

This commit adds them, although the actual way they're constructed into a DotGrammar is a bit hacky.  We need the tape name of the RegexHeader's child... but unlike with literals we can't just ask for `child.toGrammar(parsedText.text)`.  (I had really only anticipated toGrammar making one sort of thing, there isn't a straightforward way to ask it to make different things depending on context.)  So to get around this limitation, we just ask the child to construct a dummy literal, then grab the name of its tape and make a DotGrammar.

In any case, this is really just a temporary hack, because much of the relevant infrastructure will be rewritten in the near future.  (Coming soon is the ability to embed symbols in regexes, which are even less compatible with prior expectations.)