Headers now use miniParserEnv
Since miniParserEnv now does error handling the same way as miniParserMonadic, it can stand in for it when parsing headers.

There's one drawback for the time being, in that the structure of reserved words in miniParserEnv makes it more difficult to have smart error messages about them.  (E.g., if you name a header "table", we used to catch that and say "Hey you can't name a tape 'table' because that's an operator."  All we can really say now is "I can't parse that as a header".)  But I think that's kind of messaging is going to become difficult anyway, as headers and ops move closer to each other syntactically in the near future.