miniParsing now unified
Moving op parsing to use miniParserEnv means that all three cell types use the same engine, so now we can delete the other two, and have miniParserEnv be the only miniParser.  That is:

  * miniParser -> (deleted)
  * miniParserMonadic -> (deleted)
  * miniParserEnv -> miniParser

Doing this slightly damages the clarity of error messages for ops, for the same reasons mentioned in the last commit.  But that's okay, the way we were doing it was a bit ad hoc anyway; maybe sometime later we can deal with unexpected reserved word error messages in a more unified, principled way.