Disentangling header/regex grammar creation
(NB: Most of the code used in this was actually already written a long time ago, it just wasn't turned on because some bugs came up and fixing those was higher priority.  All of the Regex.toGrammar() code was actually unreachable; now it's all used.)

Turning a header-concent cell into an actual grammar was a pretty complicated process, since both could be complex, and had complicated relationships about scope.  (E.g., in something like `equals re text: a|b`, that equals scopes over the alternation.  But in something like `re embed: VERB|NOUN` or `re text/gloss: a|b` the slash actually scopes underneath the alternation.)  We basically handled this by a special function RegexHeader.toGrammarPiece(regex), which walks the parsed regex, and then once it hits a leaf like `a` or `VERB` it passes that to the header's .child (which might be a TapeNameHeader, and EmbedHeader, or a SlashHeader).

This works, and has worked for a long time, but it's tough to understand and modify, and it's in service of syntactic constructs that I'm not sure we actually want.

  * I think we actually turned of `re T1/T2: X` headers some time ago, because it's much more likely a mistake than a valid usage.  I think programmers are likely to think that means both T1 and T2 are going to have the same content... but they don't.  All that means is both T1 and T2 match X somehow, not that they're the same.  In any case, I can't really think of a valid usage of this construct, it's hard to reason about, and it's hard to support.

  * `re embed: X' has various valid usages, but it's VERY hard to reason about, can have some unexpected effects (when X has multiple tapes), and I've mostly seen programmers either ignore it or misinterpret what it meant.  I think it'll be better to support something like `re text: {VOWEL}` instead, where VOWEL is an embedded symbol with only `text` exposed.  This is more limited in some ways, but more expressive in others, it'll let us say things like `starts re text: r{VOWEL}` which were really cumbersome to express before.

In any case, if we get rid of both of those, we've got a much simpler relationship.  In something like `equals re text: (X|Y)*`, we no longer have to push `text` down to all those leaves to make the literal `text:X` etc.  There's no longer a decision to be made down there about whether that X is really a literal, embedded symbol, or slash.  We know it's a literal, and the whole regex is literals on the same tape.  When that's true, we don't need to know what tape it's on, we can just say "Call that tape whatever and rename it at a higher scope once you know what the rest of the program calls it."

Anyway, this change means that the header grammars now always scope outside of the regex grammars.  Consistent scope disentangles the construction process, so we can get rid of the awkward RegexHeader.toGrammarPiece() logic.

The only actual tricky bit getting this to work was that adjustFilters didn't expect renames inside of filters, and wouldn't push the filter through a rename, so I had to add that.